+++ src/wp/test/xp/all_test.h
@@ -49,6 +49,8 @@
 #undef TFSUITE
 #include "src/text/fmt/xp/t/fl_TableStyles.t.cpp"
 #undef TFSUITE
+#include "src/wp/ap/xp/t/ap_KeyBindings.t.cpp"
+#undef TFSUITE
 #include "src/wp/impexp/xp/t/ut_abwncrypt.t.cpp"
 #undef TFSUITE
 #include "src/wp/impexp/xp/t/ie_abinova.t.cpp"